Eliminated reporting of bogus temps
Any time there's an issue (which seems to happen a lot if on USB
to my computer connected to the mains, it will report an impossibly
high temp (basically a full int). Don't want this to mess up graphs
or other functions of the thermostat, so I've changed it to ignore
anything higher than 500C (which is the limit of the probe anyway).

I also created a variable to hold the delay time because it's now
in more than one spot and I don't want to have to hunt every instance
of it if I want to change it.